requirements.txt
appdirs==1.4.4
APScheduler==3.7.0
asgiref==3.4.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
backports.entry-points-selectable==1.1.0
boto3==1.18.16
botocore==1.21.16
cement==3.0.4
certifi==2021.5.30
cffi==1.14.6
channels==3.0.4
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.6
django-bootstrap-form==3.4
django-crispy-forms==1.12.0
django-filter==2.4.0
django-heroku==0.3.1
django-paginator==0.2a3
django-storages==1.11.1
docutils==0.17.1
filelock==3.0.12
future==0.18.2
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
jmespath==0.10.0
pathspec==0.9.0
Pillow==8.3.1
plan==0.5
platformdirs==2.2.0
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
pypiwin32==223
python-dateutil==2.8.2
pytz==2021.1
pywin32==301
PyYAML==5.4.1
requests==2.26.0
s3transfer==0.5.0
schedule==1.1.0
semantic-version==2.8.5
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
termcolor==1.1.0
Twisted==21.7.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.6
virtualenv==20.7.0
wcwidth==0.2.5
whitenoise==5.3.0
zope.interface==5.4.0